+++ Source/App/EraseInstallUITests/EraseInstallUITests.swift
@@ -1,0 +1,31 @@
+//
+//  EraseInstallUITests.swift
+//  EraseInstallUITests
+//
+//  Created by Arnold Nefkens on 17/10/2018.
+//  Copyright © 2018 Pro Warehouse. All rights reserved.
+//
+
+import XCTest
+
+class EraseInstallUITests: XCTestCase {
+
+    override func setUp() {
+        // Put setup code here. This method is called before the invocation of each test method in the class.
+
+        // In UI tests it is usually best to stop immediately when a failure occurs.
+        continueAfterFailure = false
+
+        XCUIApplication().launch()
+    }
+
+    override func tearDown() {
+        // Put teardown code here. This method is called after the invocation of each test method in the class.
+    }
+
+    func testExample() {
+        // Use recording to get started writing UI tests.
+        // Use XCTAssert and related functions to verify your tests produce the correct results.
+    }
+
+}